Move indicators to separate file
